+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.1 - 22 Jun 2025
 Improvement:
 - Export `main`, `module`, and `types` in package.json
+- Exclude `react-jsx-runtime` from `dist`
 
 Fix:
 - remove duplicated CSS